+++ .github/workflows/web_tests.yml
@@ -36,7 +36,7 @@ jobs:
 
             - name: Prepare .env
               if: steps.setup.outputs.has-changes == 'true'
-              uses: ./.github/actions/prepare-dotenv-and-types
+              uses: ./.github/actions/prepare-dotenv
 
             - name: Install Playwright browsers
               if: steps.setup.outputs.has-changes == 'true'
